Исправление багов, недочетов

+++ DiscordBot/Modules/ContentManaging/ContentModule.cs
@@ -1,6 +1,5 @@
 ﻿using Discord;
 using Discord.WebSocket;
-using DiscordBot.GuildManaging;
 using System;
 using System.Threading.Tasks;
 using DiscordBot.FileWorking;
@@ -25,59 +24,16 @@ namespace DiscordBot.Modules.ContentManaging
 
         private async Task MessageUpdated(Cacheable<IMessage, ulong> arg1, SocketMessage arg2, ISocketMessageChannel arg3)
         {
-            if (!((arg2 as SocketUserMessage).Author as SocketGuildUser).IsBot)
-                if (FilesProvider.GetGuild(((arg2 as SocketUserMessage).Author as SocketGuildUser).Guild).CheckingContent)
-                {
-                    var messUser = arg2.Author as SocketGuildUser;
-                    GuildProvider provider = new GuildProvider(messUser.Guild);
-                    var VideosAndPicturesChannel = provider.VideosTextChannel();
-                    var LinksChannel = provider.LinksTextChannel();
-                    var message = arg2;
-
-                    if (Uri.IsWellFormedUriString(message.Content, UriKind.RelativeOrAbsolute))
-                    {
-                        Uri uri = new Uri(message.Content);
-                        string host = uri.Host;
-                        bool Video = false;
-                        bool DiscordMedia = false;
-                        bool InVideoChannel = false;
-                        bool InLinksChannel = false;
-
-                        if (message.Channel == VideosAndPicturesChannel)
-                            InVideoChannel = true;
-                        else if (message.Channel == LinksChannel)
-                            InLinksChannel = true;
-
-                        if (host == "www.youtube.com" || host == "www.youtu.be")
-                            Video = true;
-                        else if (host.ToLower().Contains("discord") || host.ToLower().Contains("tenor"))
-                            DiscordMedia = true;
-
-                        if (!DiscordMedia)
-                        {
-                            if (!((InVideoChannel && Video) || (InLinksChannel && !Video)))
-                            {
-                                await message.DeleteAsync();
-                                string link = message.Content;
-                                switch (Video)
-                                {
-                                    case true:
-                                        await VideosAndPicturesChannel.SendMessageAsync(link);
-                                        break;
-                                    case false:
-                                        await LinksChannel.SendMessageAsync(link);
-                                        break;
-                                }
-                            }
-                        }
-                    }
-                }            
+            await CheckContent(arg2);
         }
 
         private async Task CheckContent(SocketMessage arg)
         {
+            var guild = ((arg as SocketUserMessage).Author as SocketGuildUser).Guild;
+            var serGuild = FilesProvider.GetGuild(guild);
+
             if (!((arg as SocketUserMessage).Author as SocketGuildUser).IsBot)
-                if (FilesProvider.GetGuild(((arg as SocketUserMessage).Author as SocketGuildUser).Guild).CheckingContent)
+                if (serGuild.CheckingContent && guild.GetTextChannel(serGuild.SystemChannels.LinksChannelId) != null && guild.GetTextChannel(serGuild.SystemChannels.VideosChannelId) != null)
                 {
                     var messUser = arg.Author as SocketGuildUser;
                     GuildProvider provider = new GuildProvider(messUser.Guild);
@@ -108,15 +64,18 @@ namespace DiscordBot.Modules.ContentManaging
                         {
                             if (!((InVideoChannel && Video) || (InLinksChannel && !Video)))
                             {
-                                await message.DeleteAsync();
+                                if((!Video && guild.GetTextChannel(serGuild.SystemChannels.LinksChannelId) != null) || (Video && guild.GetTextChannel(serGuild.SystemChannels.VideosChannelId) != null))
+                                    await message.DeleteAsync();
                                 string link = message.Content;
                                 switch (Video)
                                 {
                                     case true:
-                                        await VideosAndPicturesChannel.SendMessageAsync(link);
+                                        if (guild.GetTextChannel(serGuild.SystemChannels.VideosChannelId) != null)
+                                            await VideosAndPicturesChannel.SendMessageAsync(link);  
                                         break;
                                     case false:
-                                        await LinksChannel.SendMessageAsync(link);
+                                        if (guild.GetTextChannel(serGuild.SystemChannels.LinksChannelId) != null)
+                                            await LinksChannel.SendMessageAsync(link);
                                         break;
                                 }
                             }
